requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.46.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-ollama==0.3.1
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.2
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.34.1
networkx==3.5
numpy==2.2.4
ollama==0.4.7
openai==1.72.0
openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
playwright==1.52.0
plotly==6.0.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
st-theme==1.2.3
streamlit==1.44.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
wheel==0.46.1
yarl==1.19.0
yfinance==0.2.55
zstandard==0.23.0
